Revert "update & move `constants`, update default `uiHighlight` to #808080"
This reverts commit 68de0c5f34004b962fffc2facfacb6158e156f81.
