--- src/constants.ts
@@ -1,5 +1,0 @@
-export const CUSTOM_ENDPOINT_KEY: string = "app-portal-custom-endpoints";
-
-export const PORTAL_ID: string = "portals";
-
-export const COMMIT_HASH: string = String(process.env.COMMIT_HASH);
